+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDStringConstants.cs
@@ -166,7 +166,10 @@ namespace UnityEngine.Rendering.HighDefinition
         public static readonly int _NumTileClusteredX = Shader.PropertyToID("_NumTileClusteredX");
         public static readonly int _NumTileClusteredY = Shader.PropertyToID("_NumTileClusteredY");
 
-        public static readonly int _IndirectLightingMultiplier = Shader.PropertyToID("_IndirectLightingMultiplier");
+        public static readonly int _IndirectDiffuseLightingMultiplier = Shader.PropertyToID("_IndirectDiffuseLightingMultiplier");
+        public static readonly int _IndirectDiffuseLightingLayers = Shader.PropertyToID("_IndirectDiffuseLightingLayers");
+        public static readonly int _ReflectionLightingMultiplier = Shader.PropertyToID("_ReflectionLightingMultiplier");
+        public static readonly int _ReflectionLightingLayers = Shader.PropertyToID("_ReflectionLightingLayers");
 
         public static readonly int g_isLogBaseBufferEnabled = Shader.PropertyToID("g_isLogBaseBufferEnabled");
         public static readonly int g_vLayeredOffsetsBuffer = Shader.PropertyToID("g_vLayeredOffsetsBuffer");